requirements.txt
click==8.0.1
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.24
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
PyYAML==5.4.1
smmap==4.0.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
